UndoEntry: Merge all UndoEntry variants to one file
